requirements.txt
openai>=1.0.0
tqdm>=4.66.0
jieba>=0.42.1
pyyaml>=6.0.0
ms-swift>=3.11.0
vllm>=0.6.0
